chore(refactor): separate module for request
